Fix broken request to fecth cluster info
Agents overview controller launched an API request to detect whether Wazuh works on cluster mode or not.
It looks like the API endpoints have changed and the requested endpoints does not provide such information
anymore. This information is used to generate the search filters for the table, so the broken request resulted
on the filtering by node_name tag not showing on the search-bar filters.

- Added a new route state resolver to 'agents' view in order to fetch cluster information ('/cluster/status/).
	modified:   SplunkAppForWazuh/appserver/static/js/config/routes/agents-states.js
	modified:   SplunkAppForWazuh/appserver/static/js/controllers/agents/agents/agentsCtrl.js