Refactor `manager.py`
Refactor the required methods from the `manager.py` API controller.
This refactor was required in order to provide the `wazuhtoken` class with the required API information.

NOTE
------
At least the API key is required in order to be able to fetch its information from the database. The whole API object
is being passed currently, as the information from the database is already obtained on the refactored methods.

This behaviour is not the best as the API entry point method is in charge of a big amount of responsabilities.
A layered architecture to split responsabilities is desirable.