Fix an issue with the token where it was depecreated by the API after editing security rules
A validation system has been added to the class handling the token. If the token is cached, it is validated first by performing a basic GET request against the API and checking the response's status code (401 means it is deprecated).

This potentially solves the issue where the App becomes unresponsible after some inactivity (probably due to the token times out), and it was needed to exit and enter the app again to get a new token.