Refactor and simplification of the Python backend
The main goal of this simplification was to extract most of the duplicated code that was widely used along the different Python files.
This code has been extracted to utility functions and classes, and not accesed and mantained on an unique spot.
Some other code was no longer needed thanks to the previous work that was done on this epic (RBAC), and for this reason, has been removed.
There is also a gret improvement on how the data is handled, as for example every request to the Wazuh API are now done on a single spot (the Wazuh_API class).
A API model has been created (Data Transport Object) with some utility methods. This model is now used instead of native Python dictionaries, which where sometimes modified (adding or deleting fields), causing a big confusion as the programmer could not be sure of what data was inside the dictionary during its propagating between functions, classes and methods.

The overall structure of the backend is now much simplier and vertical, and holds its foucs on re-usablity and delegation.

Below, there is a list of changes as a result of this refactoring process:

New files
----------
- API_model.py
  - DTO object to hold the API information (address, port, user...), to replace the use of native dicts.
  - Provides some utility methods, as a result of the extraction of duplicated code.
- get_api_by_id.py
  - Utility function to get an API from the database given its ID. Returns an API DTO.
- params.py
 - Utility function to extract a value from a dicitionary given its key, if it exists.
 - Raises a KeyError exception if not.

Modified files
------------------
- api.py
 - Every request is now delegated to the Wazuh_API class.
 - Makes use of get_api_by_id and get_params.
 - Makes use of the API DTO.
 - Duplicated and no longer needed code was removed.
- manager.py
 - Every request is now delegated to the Wazuh_API class.
 - Makes use of get_api_by_id and get_params.
 - Makes use of the API DTO.
 - Duplicated and no longer needed code was removed.
 - Some endpoints could not be refactored to make use of the API DTO as it would changed its interface and would require a frontend refactoring too.
 - check_daemons.py
  - Previously extracted from duplicated code on api.py and managger.py
  - Makes use of the API DTO.
  - Every request is now delegated to the Wazuh_API class.
   - Duplicated and no longer needed code was removed.
- check_queue.py
 - Every request is now delegated to the Wazuh_API class.
 - Makes use of get_api_by_id and get_params.
 - Makes use of the API DTO.
 - Duplicated and no longer needed code was removed.
- get_agents_status.py
  - Makes use of the API DTO.
  - Refactor the requests to make use of the DTO.
  - The requests COULD be delegated to the Wazuh_API class on the future.
- Wazuh_API.py
 - Makes now use of the API DTO
 - Remove methods arguments that were no longer needed, as that information is arriving in the DTO.
- WazuhToken.py
 - Makes now use of the API DTO.
 - Remove code no longer needed.
- csvRequestService.js
 - Change the way the endpoint is named for consistency. (was 'path').