Fix an issue where the `wazuh`user coud not be used
Wrong data types were causing some conditionals to evaluate as `true` on the backend, falling to use the AUTH_CONTEXT_LOGIN instead of the regular login for the `wazuh` user.
Casting these values to the proper type (boolean) on the constructor solved the issue.
Also, an additional check has been added to prevent the `wazuh` user from using the AUTH_CONTEXT_LOGIN. Now, no other use than `wazuh-wui` is allowed to use this login method, even if the `run_as` option is enabled for the manager on the frontend (where the use of this option is explained).
This validation prevents the incorrect behaviour of the application, and will fall back to a regular login method.