Include model factory and adjust the solution
This commit includes a new set of models for the API responses, as well as a factory method to automatically create the correct model for each response. The proposed solution for the issue #1331 has been adjusted to make use of this utilities, without changing its behaviour. Also, a new function has been created on the requestService, to further simplify the workflow. This new function performs API requests, delegating on the well known apiReq() function, with the only difference that it uses the new model factory, returning a model instead of a primitive dictionary.