Refactor of the API console
This commit solves an issue on the API console, where the request payload
won't be sent using the correct content type, causing the request to fail.

In order to solve this, a new way to send requests has been implemented,
which consists on using a hierarchy of classes to model each kind of
request and response expected. With this approach, we ensure that the
request is valid before sending it. In the same way, the response is
also modeled, ensuring that the properties we need exists and are in
a valid state (this is what class constructors are for), getting rid
of errors related to the reading of undefined properties, and repeated
validation of the response object. This is all centralized now.

In order to better solve this issue, I needed to split the monstrous
Angular controller into smaller pieces, in order to better understand
its logical structure and flow. In direct benefit, much of this flow
and logic has been re-implemented using a simpler approach, and lots
of refactorings has been done, improving the code legibility and
understandability.

- More descriptive names were used for class members, variables
and methods.
- The controller has been split into smaller pieces, making it
easier to read and maintain.
- Every method has been properly documented.